requirements.txt
aiohttp~=3.9.2
Pillow~=10.2.0
asyncpg~=0.29.0
discord.py~=2.3.2